chore(content): Moves releases into directory
